+++ src/buttons/daily/dailyInterval.js
