+++ src/buttons/daily/dailyType.js
