+++ src/events/guild/guildDelete.js
