+++ src/util/buttonHandler.js
@@ -1,4 +1,4 @@
-const { readdirSync, statSync } = require("fs");
+const fs = require("fs");
 const path = require("path");
 const { Collection } = require("discord.js");
 const { ChalkAdvanced } = require("chalk-advanced");
@@ -14,27 +14,25 @@ module.exports = class ButtonHandler {
    */
   load() {
     this.loadFromPath(path.join(__dirname, "../buttons/"));
+    console.log(
+      `${ChalkAdvanced.white("Would You?")} ${ChalkAdvanced.gray(
+        ">",
+      )} ${ChalkAdvanced.green("Successfully loaded buttons")}`,
+    );
   }
 
   loadFromPath(dir) {
-    const files = readdirSync(dir);
-
-    for (const file of files) {
+    fs.readdirSync(dir).forEach((file) => {
       const filePath = path.join(dir, file);
-      const stat = statSync(filePath);
+      const stat = fs.statSync(filePath);
+
       if (stat.isDirectory()) {
-        loadFromPath(path.join(dir, filePath));
-      } else {
-        const button = require(`${filePath}`);
+        this.loadFromPath(filePath);
+      } else if (file.endsWith(".js")) {
+        const button = require(filePath);
         this.c.buttons.set(button.data.name, button);
       }
-    }
-
-    console.log(
-      `${ChalkAdvanced.white("Would You?")} ${ChalkAdvanced.gray(
-        ">"
-      )} ${ChalkAdvanced.green("Successfully loaded buttons")}`
-    );
+    });
   }
 
   /**
